requirements.txt
python-dotenv
xarray
pandas
pysofar
netCDF4
requests
setuptools
boto3
awscli
pytest
openpyxl
ioos_qc
polars
setuptools
dask
dask[distributed]
pyproj
geopy
pyarrow
geopandas
matplotlib